+++ app/src/main/java/com/playtogether_android/app/presentation/ui/home/adapter/HomeHotAdapter.kt
@@ -12,7 +12,8 @@ import com.playtogether_android.app.util.ListComparator
 import com.playtogether_android.app.util.stringListBuilder
 import com.playtogether_android.domain.model.light.CategoryData
 
-class HomeHotAdapter : ListAdapter<CategoryData, HomeHotAdapter.ViewHolder>(ListComparator<CategoryData>()) {
+class HomeHotAdapter :
+    ListAdapter<CategoryData, HomeHotAdapter.ViewHolder>(ListComparator<CategoryData>()) {
     inner class ViewHolder(private val binding: ItemHomeHotBinding) :
         RecyclerView.ViewHolder(binding.root) {
         fun onBind(item: CategoryData) {
@@ -45,21 +46,4 @@ class HomeHotAdapter : ListAdapter<CategoryData, HomeHotAdapter.ViewHolder>(List
             }
         }
     }
-
-    class HomeComparator : DiffUtil.ItemCallback<CategoryData>() {
-        override fun areItemsTheSame(
-            oldItem: CategoryData,
-            newItem: CategoryData
-        ): Boolean {
-            return oldItem.lightId == newItem.lightId
-        }
-
-        override fun areContentsTheSame(
-            oldItem: CategoryData,
-            newItem: CategoryData
-        ): Boolean {
-            return oldItem == newItem
-        }
-    }
-
 }
