+++ app/src/main/java/com/playtogether_android/app/presentation/ui/home/view/HomeFragment.kt
@@ -12,11 +12,8 @@ import com.playtogether_android.app.presentation.ui.home.adapter.HomeHotAdapter
 import com.playtogether_android.app.presentation.ui.home.adapter.HomeNewAdapter
 import com.playtogether_android.app.presentation.ui.home.viewmodel.HomeViewModel
 import com.playtogether_android.app.presentation.ui.thunder.list.view.ThunderListActivity
-import com.playtogether_android.app.presentation.ui.thunder.list.viewmodel.ThunderListViewModel
 import com.playtogether_android.app.util.viewPagerAnimation
 import dagger.hilt.android.AndroidEntryPoint
-import androidx.activity.viewModels
-import androidx.fragment.app.viewModels
 
 @AndroidEntryPoint
 class HomeFragment : BaseFragment<FragmentHomeBinding>(R.layout.fragment_home) {
