--- app/src/main/java/com/playtogether_android/app/util/ListAdapterComparator.kt
@@ -1,16 +1,0 @@
-package com.playtogether_android.app.util
-
-import android.annotation.SuppressLint
-import androidx.recyclerview.widget.DiffUtil
-
-class ListComparator<T : Any> : DiffUtil.ItemCallback<T>() {
-    override fun areItemsTheSame(oldItem: T, newItem: T): Boolean {
-        return oldItem.hashCode() == newItem.hashCode()
-    }
-
-    @SuppressLint("DiffUtilEquals")
-    override fun areContentsTheSame(oldItem: T, newItem: T): Boolean {
-        return oldItem == newItem
-    }
-
-}
