+++ app/src/main/java/com/playtogether_android/app/util/ListComparator.kt
@@ -1,15 +1,16 @@
 package com.playtogether_android.app.util
 
+import android.annotation.SuppressLint
 import androidx.recyclerview.widget.DiffUtil
-import com.playtogether_android.domain.model.light.CategoryData
 
-
-class ListComparator : DiffUtil.ItemCallback<CategoryData>() {
-    override fun areItemsTheSame(oldItem: CategoryData, newItem: CategoryData): Boolean {
-        return oldItem.lightId == newItem.lightId
+class ListComparator<T : Any> : DiffUtil.ItemCallback<T>() {
+    override fun areItemsTheSame(oldItem: T, newItem: T): Boolean {
+        return oldItem.hashCode() == newItem.hashCode()
     }
 
-    override fun areContentsTheSame(oldItem: CategoryData, newItem: CategoryData): Boolean {
+    @SuppressLint("DiffUtilEquals")
+    override fun areContentsTheSame(oldItem: T, newItem: T): Boolean {
         return oldItem == newItem
     }
+
 }
