+++ app/src/main/res/drawable/bg_selector_tab_layout.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/bg_tab_selected" android:state_selected="true" />
+    <item android:drawable="@drawable/bg_tab_unselected" android:state_selected="false" />
+</selector>
