+++ app/src/main/res/layout/activity_thunder_list.xml
@@ -104,33 +104,45 @@
             app:layout_constraintStart_toEndOf="@id/tv_thunderlist_title"
             app:layout_constraintTop_toTopOf="@id/tv_thunderlist_title" />
 
-        <LinearLayout
-            android:id="@+id/ll_thunder_align_container"
+        <com.google.android.material.tabs.TabLayout
+            android:id="@+id/tl_thunderlist_tab_container"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="24dp"
             android:layout_marginEnd="24dp"
-            android:gravity="center"
-            android:orientation="horizontal"
             app:layout_constraintBottom_toBottomOf="@id/tv_thunderlist_category_title"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toBottomOf="@id/cl_thunderlist_tool_container">
 
-            <ImageView
-                android:id="@+id/iv_thunderlist_align"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:src="@drawable/ic_icn_align" />
+        </com.google.android.material.tabs.TabLayout>
 
-            <TextView
-                android:id="@+id/tv_thunderlist_sort_type"
-                style="@style/thunder_align"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:text="@string/thunder_list_align" />
+        <!--        <LinearLayout-->
+        <!--            android:id="@+id/ll_thunder_align_container"-->
+        <!--            android:layout_width="wrap_content"-->
+        <!--            android:layout_height="wrap_content"-->
+        <!--            android:layout_marginTop="24dp"-->
+        <!--            android:layout_marginEnd="24dp"-->
+        <!--            android:gravity="center"-->
+        <!--            android:orientation="horizontal"-->
+        <!--            app:layout_constraintBottom_toBottomOf="@id/tv_thunderlist_category_title"-->
+        <!--            app:layout_constraintEnd_toEndOf="parent"-->
+        <!--            app:layout_constraintTop_toBottomOf="@id/cl_thunderlist_tool_container">-->
 
-        </LinearLayout>
+        <!--            <ImageView-->
+        <!--                android:id="@+id/iv_thunderlist_align"-->
+        <!--                android:layout_width="wrap_content"-->
+        <!--                android:layout_height="wrap_content"-->
+        <!--                android:src="@drawable/ic_icn_align" />-->
+
+        <!--            <TextView-->
+        <!--                android:id="@+id/tv_thunderlist_sort_type"-->
+        <!--                style="@style/thunder_align"-->
+        <!--                android:layout_width="wrap_content"-->
+        <!--                android:layout_height="wrap_content"-->
+        <!--                android:layout_gravity="center"-->
+        <!--                android:text="@string/thunder_list_align" />-->
+
+        <!--        </LinearLayout>-->
 
         <androidx.viewpager2.widget.ViewPager2
             android:id="@+id/vp_thunderlist_container"
@@ -141,15 +153,6 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintTop_toBottomOf="@id/tv_thunderlist_category_title" />
 
-        <!--        <androidx.recyclerview.widget.RecyclerView-->
-        <!--            android:id="@+id/rv_thunderlist_container"-->
-        <!--            android:layout_width="match_parent"-->
-        <!--            android:layout_height="0dp"-->
-        <!--            android:layout_marginHorizontal="20dp"-->
-        <!--            android:layout_marginTop="24dp"-->
-        <!--            app:layout_constraintBottom_toBottomOf="parent"-->
-        <!--            app:layout_constraintTop_toBottomOf="@id/ll_thunder_align_container"-->
-        <!--            tools:listitem="@layout/item_thunder_list" />-->
 
         <androidx.appcompat.widget.AppCompatButton
             android:id="@+id/fab_thunderlist"
